requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cloudinary==1.44.1
colorama==0.4.6
coverage==7.10.6
dj-database-url==3.0.1
Django==5.2.5
django-cloudinary-storage==0.3.0
django-filter==25.1
djangorestframework==3.16.1
gunicorn==23.0.0
h11==0.16.0
idna==3.10
iniconfig==2.1.0
mysqlclient==2.2.7
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
pytest-sugar==1.1.1
python-dotenv==1.1.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
termcolor==3.1.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
whitenoise==6.9.0
